requirements.txt
Django==1.9
lxml==3.5.0
djangorestframework==3.4.4
Markdown==2.6.6
flake8==3.0.4
xlrd==1.0.0
MySQL-python==1.2.5